refactor FeedPage out of Home
